Implement working logic for moving tiles to the board
Added extra private attributes, methods and UI updaters, so that when the player who is in the current turn has clicked a tile from their hand, they can move it to the board, it gets removed from the hand, and then they can keep doing that until they run out of tiles.

Also added a method that links the board validation to a button on the main UI, which succesfully restores the board and the player's hand to the previous state if their modifications were incorrect.